Swap H for H to toggle HTML in mu4e-view mode
h is a critical key in navigating ( `hjkl` ) and having it used to toggle HTML isn't as useful. I've swapped `h` for `H` which is a less commonly used key than `h`, so I think it should be in the default keybinding. 

This addresses https://github.com/JorisE/evil-mu4e/issues/5